commonize code for parsing comma-separated annotations
